Convert styles and materialize to scss
Instead of directly requiring the materialize css files via CDN,
they are now imported using the materialize-sass rails gem. The
new color scheme is being defined in colors.css.scss and the custom
css/scss was moved to custom.css.scss. The colors on the homepage
are updated to use the new color scheme as defined in colors.css.scss.

All original functionality from the homepage is still working, but the
style sheets are rearranged.