Create category dropdown in navbar
The desktop navbar now has a dropdown for "Jump to Category" that is
available from any page. The list of categories was created using a
helper method in ApplicationController. Additionally, I realized that
we were double-loading jQuery and materialize's javascript files
in the _javascripts partial, but they were already being loaded in
via "sprockets" in application.js. This was causing the navbar to
"flicker" when it tried to load, making it unusable. Removing the
duplication fixed this issue.