Add Model Tests for Bid Class
Adds basic model tests for bid class. Puts a character range on
the bid details, from 35 - 400 characters so messages are neither
too short nor too long. Also updates the factories so that they
conform to the bid model. Seed file has been updated for the usage
of duration_estimate as a enum (short, medium, long, event) instead
of an arbitrary number of hours.