Revert bid duration_estimate to # of days integer
We mistakenly converted the Bid class's duration_estimate to be
an enum similar to the Job class's duration_estimate, but this wasn't
what we really wanted. This branch converts it back to an integer,
where we have agreed that the integer will stand for the number of
days the contractor estimates it will take them to complete the job.
Validations and tests were added for this.